requirements.txt
EdgeGPT
openai
semaphore-bot
